New homePage
- old homePage as endpointTest
- RoomSelection as its own component to be able to reuse it in homePage
